requirements.txt
numpy
pandas
pymongo
certifi
pymongo[srv]
python-dotenv
scikit-learn
pyaml
mlflow